Fix warnings in gcc 5.4, C++11 mode
auto_ptr is deprecated in C++11, provide an alternative to
the ancient version which uses unique_ptr with a custom deleter.

+++ ImageIO/ImageSoA.cpp
@@ -79,7 +79,7 @@ struct CopyFunctor
         const int endSSE   = range.end() & ~0x3;
         assert((endSSE - beginSSE) % 4 == 0);
         assert((beginSSE-range.begin())+(endSSE-beginSSE)+(range.end()-endSSE)\
-            ==range.size());
+            == static_cast<int>(range.size()));
 
         float * r = m_args.dest.GetDataPointer<pcg::RGBAImageSoA::R>();
         float * g = m_args.dest.GetDataPointer<pcg::RGBAImageSoA::G>();
